+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 History of **rest2go** library with changes description.
 
+## 1.1.0 (2025-12-03)
+
+- `FEATURE` Handle any expected HTTP error as unknown
+- `FEATURE` Handle expected HTTP 500 error
+- `FIX` Not public HTTP server implementation
+
 ## 1.0.0 (2025-11-03)
 
 - `FEATURE` Health check